Настройка линтеров
