+++ .prettierignore
@@ -1,0 +1,4 @@
+// Любая директория с файлом App.tsx
+
+.prettierrc
+package.json
