+++ .stylelintignore
@@ -1,0 +1,7 @@
+*.tsx
+*.jsx
+*.ts
+*.js
+*.md
+*.svg
+/node_modules
