+++ .vscode/extensions.json
@@ -1,0 +1,16 @@
+// extension.json
+
+{
+    // Рекомендованные расширения
+    "recommendations": [
+      "mikestead.dotenv",
+      "streetsidesoftware.code-spell-checker",
+      "streetsidesoftware.code-spell-checker-russian",
+      "dsznajder.es7-react-js-snippets",
+      "dbaeumer.vscode-eslint",
+      "esbenp.prettier-vscode"
+    ],
+  
+    // Нежелательные расширения
+    "unwantedRecommendations": []
+  }
