+++ src/app/store.js
@@ -1,8 +1,5 @@
-import { configureStore } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
+import { configureStore } from '@reduxjs/toolkit'
 
-export const store = configureStore({
-  reducer: {
-    counter: counterReducer,
-  },
-});
+// export const store = configureStore({
+//   reducer: {},
+// })
